+++ lib/recording_studio_support/engine.rb
@@ -135,5 +135,15 @@ module RecordingStudioSupport
         FlatPack::PageNav::Component.prepend(PageNavCompat)
       end
     end
+
+    initializer "recording_studio_support.assets" do |app|
+      app.config.assets.paths << root.join("app/javascript") if app.config.respond_to?(:assets)
+    end
+
+    initializer "recording_studio_support.importmap", before: "importmap" do |app|
+      next unless app.config.respond_to?(:importmap)
+
+      app.config.importmap.paths << root.join("config/importmap.rb")
+    end
   end
 end